Cleanup filenames and file divisions.
